+++ x/delegation/keeper/keeper_suite_test.go
@@ -4,6 +4,8 @@ import (
 	"fmt"
 	"testing"
 
+	"cosmossdk.io/math"
+
 	mintTypes "github.com/cosmos/cosmos-sdk/x/mint/types"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -41,7 +43,7 @@ func CreatePool(s *i.KeeperTestSuite) {
 		Config:               "ar://DgdB-2hLrxjhyEEbCML__dgZN5_uS7T6Z5XDkaFh3P0",
 		StartKey:             "0",
 		UploadInterval:       60,
-		InflationShareWeight: 10_000,
+		InflationShareWeight: math.LegacyNewDec(10_000),
 		MinDelegation:        100 * i.KYVE,
 		MaxBundleSize:        100,
 		Version:              "0.0.0",