+++ CHANGELOG.md
@@ -29,6 +29,7 @@ An '!' indicates a state machine breaking change.
 
 ### Improvements
 
+- ! (`x/pool`) [#190](https://github.com/KYVENetwork/chain/pull/190) Make inflation-share-weight a decimal.
 - [#182](https://github.com/KYVENetwork/chain/pull/182) Make release builds reproducible.
 - ! [#183](https://github.com/KYVENetwork/chain/pull/183) Only charge coins which are whitelisted.
 - ! (deps) [#174](https://github.com/KYVENetwork/chain/pull/174) Add mainnet KYVE image to interchain tests.
@@ -38,6 +39,7 @@ An '!' indicates a state machine breaking change.
 
 ### Bug Fixes
 
+- [#192](https://github.com/KYVENetwork/chain/pull/192) fix: Cosmos-SDK & IBC swagger-code generation.
 - (`x/query`) [#159](https://github.com/KYVENetwork/chain/pull/159) Add pool params back to kyve params query.
 - (`x/query`) [#160](https://github.com/KYVENetwork/chain/pull/160) Add funders back to pool query.
 - [#163](https://github.com/KYVENetwork/chain/pull/163) Add back swagger UI for KYVE and Cosmos REST queries.